+++ src/app/api/firefly/metrics/route.ts
@@ -22,7 +22,7 @@ import { TwitterSession } from '@/providers/twitter/Session.js';
 import { TwitterSessionPayload } from '@/providers/twitter/SessionPayload.js';
 import type { Session } from '@/providers/types/Session.js';
 import { SessionType } from '@/providers/types/SocialMedia.js';
-import { getPublicKeyInHex } from '@/services/ed25519.js';
+import { getPublicKeyInHex } from '@/helpers/ed25519.js';
 
 const CryptoUsageSchema = z.union([
     z.object({
