+++ src/helpers/ed25519.ts
